refactor(backend): registry-ize InspectionService rule dispatch
Replace the 15-branch switch in InspectionService.executeRule with a
strategy + registry:
- InspectionRuleHandler strategy interface (ruleType + check)
- InspectionRuleRegistry: Spring-collects handlers, fail-fast on dup
  ruleType, dispatch unknown -> warn + empty (preserves old default)
- InspectionSupport: shared helpers (issue build/insert, rule-config
  parse, table-scope filter, view/lineage/cluster resolution, byte fmt)
- 15 *RuleHandler components, each a verbatim migration of one checkXxx
  plus its rule-specific helpers

InspectionService keeps the inspection entry, fix/fix-plan and query
APIs, delegating shared helpers to InspectionSupport (2032 -> 650 lines).
Behavior-preserving; no REST/schema change.

Tests: InspectionRuleRegistryTest (dispatch/dup/unknown) +
TableNaming/TableOwner handler tests.
